Normalement c'est finit monsieur H, je touche plus :>

+++ devices.h
@@ -10,7 +10,7 @@ typedef struct frame
     uint8_t sof;
     uint64_t destination;
     uint64_t source;
-    uint8_t type;
+    uint16_t type;
     uint8_t *data;
     uint8_t fcs[4];
 }frame;
@@ -43,7 +43,7 @@ int power(int nb, int expo);
 uint32_t build_ip(char* addr);
 int know_destintaion(bridge my_bridge, uint64_t mac);
 int compare_mac(uint64_t mac1, uint64_t mac2);
-void create_frame(frame *message,uint64_t src, uint64_t dest, uint8_t const *data, size_t size);
+void create_frame(frame *message,uint64_t src, uint64_t dest, uint8_t const *data, uint16_t size);
 void show_frame(frame *message);
 void show_frame_for_dev(frame *message);
 void print_commutation_table(bridge my_bridge);
